+++ Presentation/OpenTgResearcherDesktop/Usings.cs
@@ -1,8 +1,8 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-global using Autofac;
 global using Autofac.Extensions.DependencyInjection;
+global using Autofac;
 global using CommunityToolkit.Mvvm.ComponentModel;
 global using CommunityToolkit.Mvvm.Input;
 global using CommunityToolkit.WinUI.UI.Animations;
@@ -90,3 +90,4 @@ global using Windows.System;
 global using Windows.UI.ViewManagement;
 global using Windows.UI;
 global using WinUIEx;
+global using ZiggyCreatures.Caching.Fusion;